rest of assQuestions: authorized flag for saveWorkingData/calculateReachedPoints

+++ Modules/TestQuestionPool/classes/class.assFileUpload.php
@@ -291,7 +291,7 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
 	 * @param boolean $returndetails (deprecated !!)
 	 * @return integer/array $points/$details (array $details is deprecated !!)
 	 */
-	public function calculateReachedPoints($active_id, $pass = NULL, $returndetails = FALSE)
+	public function calculateReachedPoints($active_id, $pass = NULL, $authorizedSolution = true, $returndetails = FALSE)
 	{
 		if( $returndetails )
 		{
@@ -508,7 +508,7 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
 	*
   * @param array Array with ID's of the file datasets
 	*/
-	protected function deleteUploadedFiles($files, $test_id, $active_id)
+	protected function deleteUploadedFiles($files, $test_id, $active_id, $authorized)
 	{
 		global $ilDB;
 		
@@ -516,9 +516,9 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
 		$active_id = null;
 		foreach ($files as $solution_id)
 		{
-			$result = $ilDB->queryF("SELECT * FROM tst_solutions WHERE solution_id = %s",
-				array("integer"),
-				array($solution_id)
+			$result = $ilDB->queryF("SELECT * FROM tst_solutions WHERE solution_id = %s AND authorized = %s",
+				array("integer", 'integer'),
+				array($solution_id, (int)$authorized)
 			);
 			if ($result->numRows() == 1)
 			{
@@ -530,9 +530,9 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
 		}
 		foreach ($files as $solution_id)
 		{
-			$affectedRows = $ilDB->manipulateF("DELETE FROM tst_solutions WHERE solution_id = %s", 
-				array("integer"),
-				array($solution_id)
+			$affectedRows = $ilDB->manipulateF("DELETE FROM tst_solutions WHERE solution_id = %s AND authorized = %s",
+				array("integer", 'integer'),
+				array($solution_id, $authorized)
 			);
 		}
 	}
@@ -618,7 +618,7 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
 	 * @param integer $pass Test pass
 	 * @return boolean $status
 	 */
-	public function saveWorkingData($active_id, $pass = NULL)
+	public function saveWorkingData($active_id, $pass = NULL, $authorized = true)
 	{
 		global $ilDB;
 		global $ilUser;
@@ -657,7 +657,7 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
 		{
 			if (is_array($_POST['deletefiles']) && count($_POST['deletefiles']) > 0)
 			{
-				$this->deleteUploadedFiles($_POST['deletefiles'], $test_id, $active_id);
+				$this->deleteUploadedFiles($_POST['deletefiles'], $test_id, $active_id, $authorized);
 			}
 			else
 			{
@@ -666,22 +666,20 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
 		}
 		elseif( $checkUploadResult )
 		{
-			if (!@file_exists($this->getFileUploadPath($test_id, $active_id))) ilUtil::makeDirParents($this->getFileUploadPath($test_id, $active_id));
+			if(!@file_exists($this->getFileUploadPath($test_id, $active_id)))
+			{
+				ilUtil::makeDirParents($this->getFileUploadPath($test_id, $active_id));
+			}
+			
 			$version = time();
 			$filename_arr = pathinfo($_FILES["upload"]["name"]);
 			$extension = $filename_arr["extension"];
 			$newfile = "file_" . $active_id . "_" . $pass . "_" . $version . "." . $extension;
+			
 			ilUtil::moveUploadedFile($_FILES["upload"]["tmp_name"], $_FILES["upload"]["name"], $this->getFileUploadPath($test_id, $active_id) . $newfile);
-			$next_id = $ilDB->nextId('tst_solutions');
-			$affectedRows = $ilDB->insert("tst_solutions", array(
-				"solution_id" => array("integer", $next_id),
-				"active_fi" => array("integer", $active_id),
-				"question_fi" => array("integer", $this->getId()),
-				"value1" => array("clob", $newfile),
-				"value2" => array("clob", $_FILES['upload']['name']),
-				"pass" => array("integer", $pass),
-				"tstamp" => array("integer", time())
-			));
+			
+			$this->saveCurrentSolution($active_id, $pass, $newfile, $_FILES['upload']['name'], $authorized);
+			
 			$entered_values = true;
 		}
